refactor: match interface name
