+++ packages/shared/src/models/service-less-quadlet.ts
